Add url.js file to apply on fetch
move Fetch functions to its own module fetch.js
